[FIX] documents: fix missing action buttons
Purpose
=======
Fix the missing action buttons when:
- uploading files
- doing a select all using CTRL+A

Specification
=============
The showActions method is responsible for the display
of the different action buttons.
The actions are currently shown only if the right panel
doesn't have any previewed document and no focused document
or a focused document which is selected.

When uploading files, the uploaded documents are selected but
there is no focused document meaning the action buttons are hidden.
Fixing that by focusing the first uploaded document when all
the uploaded documents have been selected.

When selecting all the documents using CTRL+A, the records are
also selected but there could be no focused document or it could
be out of the selection meaning the action buttons are hidden.
Fixing the issue by making sure that, in those cases, the focused
document is the first of the selection.

related commit: odoo/enterprise@d6d3d3fd07c6174bd8c5a84a420876415f2b271a

Task-4737797

closes odoo/enterprise#85821

X-original-commit: effab12888219c96a239357ec217f273a34be0f2
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>