[IMP] l10n_be_codabox: update button class for CodaBox fetch actions
The "Fetch from CodaBox" link displayed on the accounting dashboard
had no gap between it and the button at the left of it, and wasn't
aligned horizontaly with the other buttons next to it making it look
out of place.

It is now displayed as a link button, which solve both problems.

task-4671468

closes odoo/enterprise#86796

X-original-commit: 031b42349a223ca836be87ac027b7325cb4dbc88
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>