[FIX] sale_commission: avoid filtering too much in details
Before this commit, when the user clicked on "details" it would put in
the context the user_id of the current commission and his team. It was
too restrictive as some record can belong to another user and match the
team or match the team and have a different user depending on the
configuration of the plan. As a result clicking on details would not
always display all the expected achievements.

Taskid: 4776256
X-original-commit: 77eeb0cb0c5e9e2d62780fca17f14738943e03bf
Part-of: odoo/enterprise#86029
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>