[FIX] hr_expense_extract: fix OCR bad completion
When an expense is manually created, the document is sent to the OCR,
and the ocr modifies the values of the document.

The issue is that the user cannot see that values have been modified and
submit the expense with the wrong values (the ones returned by OCR which
are not displayed on the form during submission).

So to avoid such behaviour, we've made different fixes:
- We display the banner that it has been sent to the OCR immediatly (not
  after a refresh)
- We make work the refresh button diplayed in the banner
- We don't change the name and/or product of the expense if the name of
  the expense has been entered manually (to know that we check if the
  name is the generated one when we create from attachment or not)
- We don't modify the status of the expense to draft (if it has been
  submited before the response of the OCR we don't want modifications)

The test have been modified the use case, because to correctly work if
the expense is sent to the OCR, it means that there is an attachment
linked to this PR.
Also when we want to check the completion of the OCR, the attachment
name is set as expense name (expense name is required)

closes odoo/enterprise#86276

X-original-commit: ecee3c8e6185a98cbf722dc9bfc239521cc9c275
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>