[FIX] marketing_automation_sms: trigger bounce on SMS
Previously to this fix, an error considered as a 'bounce' for SMS did
not trigger the 'sms_bounce' event for linked marketing activities.
It means 'on sms bounce' activities are never fired as we only updated
mailing traces, not marketing traces.

Task-4066243: [marketing_automation] Rescheduling issues

X-original-commit: 45ebfd2026c0fe2fa362f156cf3861dbdaa7db74
Part-of: odoo/enterprise#85247
Related: odoo/odoo#209409
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>