[IMP] l10n_br_edi_pos: add the ability to send errored NFC-e orders in batch
Brazil's EDI servers are known to go down for hours or days on end, so
companies will continue to sell goods and eventually send all of the
invoices off to the servers when return back online.

Right now, the only way to do that for a POS order is to go to them one
by one and hit Send NFC-e. This PR adds a new action to the list view of
POS Orders that will mark orders as ready to be sent off to the EDI
servers. As each order takes a chunk of time to send, this job is passed
off to a new scheduled action based off of the Send and Print batch
wizard.

Records will be processed in batches and once they are completed the
cron will notify the user that originally triggered it to let them know
if everything was successful or if there are new errors to investigate.

task-4688982

closes odoo/enterprise#83840

Signed-off-by: Laurent Smet (las) <las@odoo.com>