[FIX] l10n_br_avatax: change pop-up title after pinging Avatax BR API
The previous connection test window showed up as an UserError with the
title "Invalid Operation" which can be confusing to users if the
connection ends up as successful similar to what was done for AR.

Replace the UserError with a RedirectWarning that has a better title
message ("Odoo Warning"). As it needs to redirect to somewhere, we
redirect back to the current accounting page in settings.

task-4772548

closes odoo/enterprise#84956

Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>