[FIX] marketing_automation_whatsapp: narrower parent search
When receiving an incoming whatsapp message, we set parent as replied.
As we don't always have parent_id set due to whatsapp limitations we
also search for whatsapp messages sent to the same number and linked
to marketing traces to set it as parent.

However when we correctly found a parent on whatsapp message we should
not do this additional search, as it may impact other traces in a given
marketing campaign, search being quite broad. We also limit the search
to the last found message, not all messages.

Task-4066243: [marketing_automation] Rescheduling issues

X-original-commit: 35055399d90853d30c78275eba7d4b744881acee
Part-of: odoo/enterprise#85247
Related: odoo/odoo#209409
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>