[FIX] account: Ignore zero balance lines in reconciliation
Description of the issue this commit addresses:

Zero balance lines are creating friction when reconciling lines. They will
never need a renconciliation but they still impair the user's ability to
reconcile other lines together if selected.

Desired behavior after this commit is merged:

When reconciling lines, the zero balance ones are simply ignored and the
reconciliation proceeds as if they were not selected.

task-4723956

closes odoo/enterprise#85111

X-original-commit: 3a854e24ff843a64c83ea83f8d2391bdda307588
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>