[FIX] l10n_be_coda: globalisation details import coda
When importing a coda statement, in case of globalisation,
Odoo doesn't take into account the different levels of details.
Instead of that, we keep every lines even if one is already
the sum of each detailed lines.

According to https://febelfin.be/media/pages/publicaties/2023/archief-bankstandaarden/0c9e6e5026-1694763197/standard-coda-2.4-fr.pdf

Chapter 3.1

Transaction type are communicate like that:

- 0 -> simple amount without detailed data
- 1 -> amount as totalised by the customer
	- 5 -> detail of 1
- 2 -> amount as totalised by the bank
	- 6 -> detail of 2, simple amount without detailed data
	- 7 -> detail of 2, simple amount with detailed data
		- 9 -> detail of 7
- 3 -> simple amount with detailed data
	- 8 -> detail of 3

Let's assume this case:

- 1 line with transaction type == 2, amount == 3000
- 1 line with transaction type == 7, amount == 2000
- 1 line with transaction type == 9, amount == 2000
- 1 line with transaction type == 7, amount == 1000
- 1 line with transaction type == 9, amount == 1000

Currently, Odoo will display ines 1, 2 and 4 in case of
split transactions instead of lines 3 and 5.
In case of not split transactions, Odoo sums up the same lines.

With this commit, we fix the way we build the globalisation stack
using the movement reference and globalisation codes instead of
movement reference only.

opw-4619660

closes odoo/enterprise#87580

X-original-commit: 090977f35d0e6f01457bc36e95b1ff39ebb0b83d
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>
Co-authored-by: svfu-odoo <svfu@odoo.com>