[FIX] account_online_sync: display data on bank/cash/credit journals
A previous commit 9b133b2cd03a726d330c87a1227903fc925b0ed5 restricted the view
and actions we had on the accounting dashboard for a bank journal from a branch
company, if the journal belongs to the parent company.

In Odoo 17 the fix made sense, but in Odoo 18, the dashboard data added
conflicts with another one, resulting in no data being displayed at all when
the journal is connected to an online account.

This PR adapts the initial fix to use the existing data (deprecate
`show_bank_connect` in favor of `display_connect_bank_in_dashboard`) and adds
the 'credit' type journal in the condition for `show_sync_action` to be used
by the dashboard data.

closes odoo/enterprise#85179

X-original-commit: 63c69a5d48932d41ef779a477ef88cbf244228a1
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florent Huylenbroeck (flhu) <flhu@odoo.com>