[FIX] documents_account: multi company error on attachment processing
Before this commit
------------------
if a user had a multi-company setup with a different alias domain
set for each company, a multi-company error could happen.
It happens when sending a mail with attachment to an accounting
journal alias.
The exact conditions required to enable the option to centralise
all attachments in Documents for the journal in question.

steps to reproduce:
1. clean DB with documents_account installed
2. create a 2nd company "company 2"
3. create an alias domain for company 1 "alias.company1.com" and another
alias domain for company 2 "alias.company2.com"
4. in documents, create a "finance" folder for company 1 (i.e. set the
company_id field on the folder) and another folder for company 2
5. configure the settings for both companies so that all attachments
are centralized in documents for accounting, in the respective folder
for that company
6. create an alias for some journals in company 2 (e.g.
vendor-bills@alias.company2.com)
7. send an email with attachment to that alias
-> you will get a multi-company error

After this commit
------------------
Sending a mail with attachment to an accounting journal alias creates
the account_move as expected. The attachment is parsed as expected.

opw-4727873
opw-4655559
opw-4746321
opw-4766895

closes odoo/enterprise#90069

X-original-commit: 7a7d7f48958269b252e9e7b1e5f31c65b7469334
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Victorien Sonneville (viso) <viso@odoo.com>