[FIX] l10n_cz_reports_2025: fix c_evid_dd in VAT Control Statement report
Steps to reproduce:
- Install l10n_cz_reports_2025
- Switch to a Czech company (e.g. CZ Company)

- Create a bill:
  * Customer: [a Czech customer]
  * Bill Reference: [any] (e.g. XYZ)
  * Invoice Lines: [a line with an amount higher than 10000 and a tax]
- Confirm the bill

- Go to "Accounting / Reporting / Statement Reports / Tax Report"
- Select "VAT Control Statement (CZ)" as Report
- Select a period covering the bill date

The bill should appear in the report under "B.2. Received taxable
supplies and provided payments above CZK 10,000" section.

Issue:
The "Tax document registration number" (i.e. "c_evid_dd" in the exported XML)
value for the bill is the name of the bill.

According to the documentation:
https://financnisprava.gov.cz/en/taxes/VAT-Control-Statement
https://adisspr.mfcr.cz/dpr/adis/idpr_pub/epo2_info/popis_struktury_detail.faces?zkratka=DPHKH1
For purchase-type documents, this field should have the bill reference
as value.
It should identify the original reference coming from the vendor.
If there is no bill reference, the field should be empty.

opw-4737630

closes odoo/enterprise#86701

X-original-commit: d9fd3aecd38a6f46b874e0181a123b8bc2a10e42
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>