[FIX] ai_fields: fix get ai m2o properties
Purpose:
--------
Following the changes in [1], the relational model uses an object instead
of an array for m2o values.

However, the "Refresh value" button next to an ai m2o property still returns
an array, which results in showing "No Access" as m2o value.
This commit changes the returned value to an object.

Task-4775956

[1]: https://github.com/odoo/odoo/pull/202534

closes odoo/enterprise#85007

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>