[PERF] documents: avoid fetching datas if unnecessary
The 'write' override could unnecessarily fetch the 'datas' field (binary file content).

This happened because a condition checking 'record.datas' was evaluated before checking if
'datas' or 'url' was actually present in the values being written. This could lead to
slowness when e.g. archiving/unarchiving documents, as the ORM would fetch the binary content
even when it was not needed.

We leverage Python's boolean short-circuiting to ensure that this behavior is avoided.

closes odoo/enterprise#88682

X-original-commit: ce298f671f5c4049a1610066d3a524ba7b698217
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>
Signed-off-by: Julien Segers (juse) <juse@odoo.com>