[FIX] project_timesheet_forecast: don't fetch slots when row fields of grid are not valid
Before this commit, when the user is in `Timesheets > To Validate > All
Timesheets` and group by `Sale Order Item`, a traceback is occurred
because no fields are given for the group by of web_read_group made on
`planning.slot`, the reason is because we just check if `project_id`
and/or `employee_id` are used in row fields or section field of the grid
view of timesheets to add additional data based on planning.slot in the
same period of the grid view.

This commit avoids calling the web_read_group on planning.slot model
when no rows managed are found in the grid view since the result of the
web_read_group will be useless in that case.

Steps to reproduce
------------------

0. Install `project_timesheet_forecast` module
1. Go to `Timesheets > To Validate > All Timesheets`
2. Go to the grid view
3. Group by `Sale Order Item`

Expected behavior
-----------------

The grid view should be loaded without any issue and show the so_line
field as row.

Current behavior
----------------

A traceback is occurred inside the web_read_group on planning.slot
model.

task-4782077

closes odoo/enterprise#85232

X-original-commit: a8a7349f5d172d1392f3425029d807b521cc837d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>