[FIX] account_accountant: Reload reconcileCountPerPartnerId
The aim of this commit is making sure that we update the
reconcileCountPerPartnerId properly.

Before this commit, the code was only updating partner id who
still have line to reconcile. It was wrong as we want to update
a partner who doesn't have a line to reconcile anymore.
In the meantime, we're now calling the function who reloads the
count when we delete a reconciled line.

task-4782264

closes odoo/enterprise#85171

X-original-commit: cff11d4b74559355ad6a1d74ab41e124dbd78d70
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>