[FIX] account_accountant: Use amount residual
This commit will replace some occurrence of the balance for the amount residual
and amount currency for the amount residual currency.

In this example:
Invoice of 100$

A first statement line of 20
A second statement line of 100

We do the reconciliation between the invoice and the first statement line.
The invoice will be partially reconciled. Before this commit since we used the
balance we wouldn't take this reconciliation into account when doing the
reconciliation between the remaining amount of the invoice and the second
statement line.

no task id

closes odoo/enterprise#85885

X-original-commit: 8bc9b75dd9a44ee87eb3b222cbec82a9697d8e6c
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>