[FIX] l10n_mx_edi_pos: prevent error when selecting company contact
Before this commit, selecting a contact of type "company" on the
payment screen would result in the following error:

`TypeError: Cannot read properties of undefined (reading 'name')`

This occurred because selecting a company contact automatically set the
order to be invoiced, even if the required fields were not filled in.

To prevent this error, the process now checks that all necessary fields
are completed before setting the order to invoice.

opw-4773831

closes odoo/enterprise#87145

X-original-commit: dc00b20c859b456a624a155609f14406b953e9b7
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>