[IMP] account_accountant: limit models to show in bank reco widget
We should only show reco models that should be applicable on the lines.
And so, checking the partner/label/... so they match the values that
are set on the reco model (or they are left empty on the model).

We also moved this part to the new service.

task-4760590

temp

X-original-commit: a5c37f7140b2663b3b652c43d7338e0585975b5b
Part-of: odoo/enterprise#85814
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>