[FIX] timesheet_grid: allow to edit project/task on timesheet linked to timer
Before this commit, the user cannot edit the timesheet in the timer
header inside the list view because the record is considered to be
readonly by the list view. The reason is because the timesheet is
considered as readonly when the timer is started which is wrong.

This commit makes sure the project, task and name fields of timesheet
are still editable even if a timer is linked to that timesheet. This
commit also makes sure the employee, unit amount and date fields are
readonly if timer is running on the timesheet.

task-4781902

closes odoo/enterprise#85248

X-original-commit: 2a777a6a692dd7a5d1159bf330160e601c40ac9d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>