[FIX] timesheet_grid: allow timer to timesheet user only
Before this commit, the project user cannot access to the form view of a
task with timesheets recorded, because the fields related to the timer
are always computed and crash if the user does not have any access to
timesheet app.

This commit makes sure the timer fields are only read if the user has
access to timesheets app.

task-4782330

closes odoo/enterprise#85235

X-original-commit: dfb52cf4c2fe764466644bdc4ffea5b42de7a5be
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>