[FIX] mrp_mps: update product with many components
When updating a parent product, the MPS view will load ALL the components
on the page even if it breaks the pager limit.
Previously, there was a refresh at every change. It was annoying so it was
removed. However this refresh made sure to reduce the number of records shown
to the pager limit.
The solution is to simply slice the received array of lines to the n
first records, n being the value of `this.env.config.limit`.

opw-4778588

closes odoo/enterprise#85281

X-original-commit: 590abb11454eee2b836b7290e09eda3052d9b8b5
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>