[FIX] helpdesk: search partner in specific company or no company set
Before this commit, the #84245 adds an extra_domain to find the customer
to assign on the ticket based on the email address received. The problem
is the partners could also have no company set and so the method could
not find the existing partner with the email given and will create a
duplicate partner because of that.

This commit adds inside the extra_domain to also search on partner
without any company set to be sure to not create duplicate the partner.

closes odoo/enterprise#85757

X-original-commit: 8e812b7944e8746223644da315ab401e64490673
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>