[FIX] sale_planning: add 2 needed fields in sale order form view
Before this commit, due to a cleaning of sale order views, the
`planning_first_sale_line_id` and `planning_initial_date` has been
removed from the sale order form view but those fields have to be loaded
in the form view to be used as default values when the user clicks on
`Planned\n x Hours` stat button, otherwise a traceback will be occurred.

This commit introduces again those fields in the sale order form view to
be sure those fields are defined and used as default value for the
`Planned\n x Hours` stat button.

Steps to reproduce
------------------

0. Install sale_planning module
1. Go to Sales app
2. Create a quotation with a product in which slots will be generated
3. Confirm the quotation
4. Click on `Planned\n x Hours` stat button

Current behavior
----------------

A traceback is occurred because `planning_first_sale_line_id` and
`planning_initial_date` fields are not defined in the form view.

Expected behavior
-----------------

The action of the stat button should be executed without any issue.

task-4781696

closes odoo/enterprise#85300

X-original-commit: 2df4909dc737cb49f22e672a272675a2bbfb0174
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>