[IMP] account_batch_payment: add constraint
For the creation of batch payment, the payment method must be the same. But i
can happen that the payment method was firstly used without an outstanding
account and so it will result in a batch payment with both entries and no
entries. This will add a constraint to avoid that behavior

task-4749334

closes odoo/enterprise#90861

X-original-commit: 9dce0f776dd9db62c919393acda034a30c66c369
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>