[FIX] helpdesk_sale_timesheet: don't use form view of SO to edit SOL
Before this commit, when the user would like to go to the form view of
SOL linked to a ticket (when he is on ticket form view and click on the
arrow right next to the sale_line_id field), a traceback is occurred
because a field is undefined in `sale.order.line` model. The reason is
because the form view loaded is the one of the SO.

This commit changes `form_view_ref` into `so_form_view_ref` because the
`so_line_create_button` could need to use another form view of SO than
the default one when the user clicks on `fa-plus` button to create a new
SO.

Steps to reproduce the issue
----------------------------

0. Install `helpdesk_sale_timesheet` module
1. Create a billable helpdesk team
2. Create a new ticket inside that helpdesk team
3. In the form view of that ticket, clicks on the `+` button displayed
    next to the SOL field to create a new SO
4. Create the SO with at least one service product in the SO and save
5. In the ticket form view, open the form view of the SOL set on the
   ticket.

Expected behavior
-----------------

The SOL form view should be opened without any issue.

Current Behavior
----------------

A traceback is occurred because a field does not exist in SOL model, the
reason is because we open the form view of SO instead of the SOL one.

task-4781247

closes odoo/enterprise#85262

Related: odoo/odoo#209446
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>