[FIX] l10n_de_pos_res_cert: prevent error when pay order in mobile menu
Before this commit, validating an order from the mobile menu caused an
error because the 'name' field was removed from the order in the
process_order method of self order. Since the 'pos_reference' field
exists and is correct to use (unlike 'name', which may be '/'), the
code has been updated to use 'pos_reference' instead.

opw-4758288

closes odoo/enterprise#87137

X-original-commit: 9525b7ad9745370e840e617e97536c5a3bf59b25
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>