[FIX] esg_project: prevent loop recursion error when deleting project
Before this commit, when the user tries to delete a project in project
app, a traceback is occurred because of recursion loop in the code
during the deletion of the record.

This commit removes the recursion in the code to correctly delete the
project without any issue.

Steps to reproduce the issue
----------------------------

0. Install `esg_project` module
1. Create a project
2. Go to Projects list view
3. Delete the project created in step 1

Expected behavior
-----------------

The project should be deleted without any issue

Current behavior
----------------

A recursion error is occurred during the deletion of the project

task-4788173

closes odoo/enterprise#86148

X-original-commit: 7fc1a36f012184010f0fa8f1cb88b868d3ec29ba
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>