[FIX] l10n_au_hr_payroll_account: avoid NewId warning in compute
steps to reproduce:
1.install l10n_au_hr_payroll_account
2. without demo data
3. run the test test_form_new_record

The compute method was being triggered too early, while the record still had a temporary NewId and not a real database ID.
This caused a warning when using the field in a domain: Domains don't support NewId.

We added a dependency on previous_report_id to make sure the compute only runs after the field is set and the record is saved.
This avoids the warning and keeps the logs clean.

build_error-115303

closes odoo/enterprise#87823

X-original-commit: ad886c040c377de8dcdf64e000e0df4f46c66f89
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Omar Sherif Ali Hassan (osah) <osah@odoo.com>