[REV] marketing_automation, web_studio: Supports dynamic domain
This reverts the commit fbfa23e. Indeed, for security reasons, it is
preferable to avoid calling safe_eval whenever possible.
Note that in https://github.com/odoo/odoo/pull/210576 a new mechanism is
introduced in order to block the creation of expressions in domain
fields where we only want domains with literals because they are evaluated
with literal_eval.

closes odoo/enterprise#86124

X-original-commit: 04eb11a51583adc05e0d70bacc0afdf9fbfee3d9
Related: odoo/odoo#211105
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>