[FIX] account_accountant: Limit models visible on wizard
Before, only button reconciliation models were visible in the manual
reconciliation wizard.
This filter was removed as the button mode got removed but we still
don't want every model to be visible.
So now, we filter the reconciliation models shown there based on the
partner and the amount.

Also removed some fields that don't exist on the reconciliation widget anymore to avoid traceback.
We will need to decide later what to do with the to_check field

task-4760590

closes odoo/enterprise#85814

X-original-commit: 3fdaeb1dbb20da7d7f7fd4a3c3b30184ac528484
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>