[FIX] timesheet_grid: don't focus timer on popover click
**Steps to reproduce**
Reproduced on Linux/Firefox. Not reproducible on all browsers.
- Open Timesheets app.
- Try to add a custom grouping via the "Add Custom Group" select.
Bug: the dropdown closes immediately

**Cause**
Commit 72af4ca4d9859e7236a21cb71401d9f690c42eb6 added an event listener
https://github.com/odoo/enterprise/blob/72af4ca4d9859e7236a21cb71401d9f690c42eb6/timesheet_grid/static/src/components/timesheet_timer_header/timesheet_timer_header.js#L41-L48
Depending on the browser, the first click on a `select` may not propagate
to its anchestors all the way to the document `body`. However, if it does,
the timer button is focused and the dropdown closes.

**Solution**
Exlude clicks on popover elments from focusing the timer Start/Stop button.

opw-4768571

closes odoo/enterprise#85668

X-original-commit: 98d654d3c9bf6a647c163eece0d3e7212d32d9b7
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>