[FIX] l10n_be_reports_prorata: xsd changed, rounding
The xsd changed. Now the rounding needs to be to 2 decimals everywhere.
The official xsd does not mention it, but the field submit_more must
appear on the xml.

Also, while it's not blocking directly, the declaration will be rejected
if the special decudction is filled without specifying the special %age.
So we add a banner warning the user of it.

opw-4789154

closes odoo/enterprise#86052

X-original-commit: ef43b52afdae020173c7573e9d6ddc57f37f18e3
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>