[FIX] account_reports: fix Trial Balance with analytic plan groupby
Steps to reproduce:
- Install Accounting
- Activate "Analytic Accounting" in Accounting settings
- Go to "Accounting / Accounting / Journal Entries"
- Create an entry with some analytic distribution
- Go to "Accounting / Reporting / Audit Reports / Trial Balance"
- Activate "Analytic Group By" in the settings of the report
- Add the plan used in the analytic distribution in "Analytic" group option

Issue:
In the "End Balance" column, all the amounts are summed in the total
column instead of being grouped by analytic plan.
It results on a total amount being twice what it should be as the
total column of each period is also included in the sum.

This is a complement to previous fix for the grouping by analytic account
where the analytic plan groupby case was forgotten:
https://github.com/odoo/enterprise/commit/05bbab713c749b7d7450d5dbaf2eda1ba4f68712

Note:
From saas-18.1, there is no issue.
Just forwarding the test.

opw-4648009

closes odoo/enterprise#85485

X-original-commit: fa9a6e59cdc4459fe8f37ba7eb7c94ee62eae720
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>