[FIX] pos_blackbox_be: log IP without request
Before this commit, when pos_blackbox_be was installed, the pos tests
were failing because most pos test called `_run_test` which is calling
`_satrt_pos_session` which is calling `open_new_session` which
is calling `set_opening_control`. This makes that when calling
`set_opening_control`, the request is not bound and we cannot do
`request.geoip.ip`.

We now check that the request is bound in `_log_ip`: if it is not,
we return, if it is, we resolve the ip on the fly.

We also add a check to the sale details data computation method
`get_sale_details` override in pos_blackbox_be to check if the
config is a blackbox one.

original commit: df796f3770ff2fdaf553cdfb4206f65c7b371189

closes odoo/enterprise#85484

X-original-commit: 59acb437167a9bce8696c1e940863186920aa379
Related: odoo/odoo#209908
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>