[REM] test_rental_product_configurators
Rental doesn't have a configurator anymore since 17, there
is no need for a specific module to test integration with other
configurators.

Furthermore, the refactorings of the configurators already introduced
clean tests in `sale_renting` (and `website_sale_renting`) modules,
so there is no need to keep the tour of this module.

closes odoo/enterprise#86415

Related: odoo/odoo#211651
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>