[FIX] account_accountant: hide redundant secondary buttons
Before this commit:
The same button could appear in both the primary and secondary button groups on
a bank statement line, leading to unnecessary duplication.

After this commit:
If a button is already shown as a primary action, it is excluded from the list
of secondary buttons to avoid redundancy.

task-4780841

closes odoo/enterprise#86244

X-original-commit: 1d9636eb5e9b319f2d47117a3fe1a11561a30374
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>