[FIX] account_budget: Lines sum not marked above budget
Description of the issue this commit addresses:

When multiple lines that are in budget on their own are above budget when
summed together, neither the lines nor the budget button show that the budget
will be exceeded. This leads to making purchase orders that shouldn't be made.

Desired behavior after this commit is merged:

This commit makes it so that when the lines together will exceed the budget,
the budget smart button and all the lines become red as a single line would.

task-4710838

closes odoo/enterprise#85553

X-original-commit: 48dfc2dfe976660b079f5c41c843244e32dfffa3
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>