[IMP] l10n_uy_edi: 'Lock Posted Entries with Hash' field visibility on UY electronic journals.
In uruguayan companies, if the field "Lock Posted Entries with Hash" is checked, the user can not validate electronic invoices in DGI (Gubernamental organism for electronic invoicing).
In this PR we are hiding the field in electronic uruguayan journals to prevent the user to check that field accidentally and generate inconsistencies with electronic invoicing.

closes odoo/enterprise#87636

X-original-commit: f81b0811f5152ebb38f6d18a27fcfbc6e851f6cd
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>