[FIX] l10n_it_reports: include “Other enterprises” calculation in Italian balance sheet
In the Italian balance sheet, the section for “d-bis. Other enterprises” is not added to the calculation for the section “1. Holdings in:”. As a result, it is not included in the calculation for assets in the balance sheet.

This commit adds the section “d-bis. Other enterprises” to the calculation of “1. Holdings in”.

This specification is based on Article 2424(1) of the Italian Civil Code:
https://www.normattiva.it/uri-res/N2Ls?urn:nir:stato:regio.decreto:1942-03-16;262

Steps to Reproduce on Runbot:

1. Install l10n_it_reports
2. Create an account starting with 1306
3. Create a journal entry using account 1306 as the debit line and balance it with a credit line
4. Check the Italian balance sheet and notice that “d-bis. Other enterprises” is not included in the calculation for assets in the Italian balance sheet

opw-4763868

closes odoo/enterprise#85912

X-original-commit: 65574a15c4cb7db29ba4242da57fec57b0f713d6
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>