[FIX] hr_payroll: use local times for payslip attendance
This commit fixes an issue where we had been using
the UTC start and end times for a work entry when
computing the correct duration for a payslip.
Specifically, this is necessary when a work entry
exceeds the interval of the payslip, which can
happen when a work entry starts before the payslip
interval or ends after the payslip interval.

Previously, the UTC times were used. Consider an
EST work entry that starts at 18:00 one night and
ends at 01:00 the next morning. The UTC times
would be 22:00 and 05:00 respecively, up to a
1-hour difference for daylight savings time.
Now suppose that the paysliip for this work entry
only goes up to that following morning. Instead of
using the local times of 18:00 and 01:00, the
payslip would use the UTC times of 22:00 and
05:00, which would only count 2 of those hours
instead of 6.

This is resolved by converting the UTC times to
local before using them to clamp the work entries
to the payslip interval.

closes odoo/enterprise#86117

X-original-commit: 09ce47395b0b615c55baf6689560eba3a37dfdee
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Harrison Hutton (hahu) <hahu@odoo.com>