[FIX] worksheet: don't delete templates on uninstall
During uninstall, fields (columns) bound to the module get deleted
first, so the columns used for the search don't exist anymore and the
search fails, which leads to the tables not being properly dropped,
which can then lead to the module reinstallation not being
clean e.g. because there are rows left in the table which can lead to
constraints not being addable on install.

This has been the cause of `resource` failing forever in the uninstall
nightly test: it's most likely been failing since this hook was introduced.

closes odoo/enterprise#85534

X-original-commit: df0be0b9296d7dea663087b425d4965380d7ab88
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>