[FIX] ai: fix missing `flex-grow-1` class on agent config page
This commit fixes an issue about the title and subtitle of an AI agent
being cut on the configuration form view.

Prior to this commit, the layout was using a regular div, which badly
computed the size of the inner content, which resulted in a cut text due
to the lack of a `flex-grow-1` class, as we do in other form views.

To handle the issue, we simply add the missing class to the layout.

task-4793748

closes odoo/enterprise#86044

X-original-commit: 78cfa6a543266bace6f3726bb83f092acd35a78a
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>