[FIX] l10n_be_reports: 281.50 only report form > 250 EUR
The aim of this commit is to:
    - avoid reporting negative amount in a 281.50 form which are
      rejected by the administration.
    - avoid reporting form with 0 value everywhere which are
      rejected by the administration.
    - avoid reporting a form with an amount < 250 EUR

Context:
The administration stated that we don't need to make any report that
would be above 250 EUR.
(art. 57, al. 3 CIR 92 et art. 30 AR/CIR 92)

https://www.wolterskluwer.com/fr-be/expert-insights/tax-form-281-50
https://travaillerensociete.indicator.be/main/newsadvice/52784/news/52785/news_story/533/
https://www.monastucesetconseils.be/2025-04/fiches-fiscales-281-50-pour-le-29-juin-2025-WAACPCAR_EU33150802

Before this commit:
Odoo could create a form with a negative amount as commission/fee/...
This could result in the administration software converting the xml into
a BOW file to generated an error and a file that will be rejected by
Belcotax.
Same for form with 0 everywhere.

Also, accountants always end up deleting the 281.50 form below 250 EUR
as those don't need to be reported.

After this commit:
Odoo don't report negative or null total amount as commission/fee/... and
doesn't create any form if the total amounts are below 250 EUR saving
accountants the burden to remove them.

task-id: None
This was discussed and done with TSB and DELC and is also the result of
our past experiences reporting 281.50 forms which also involved CAHE,
MAKR and CAL.

closes odoo/enterprise#85716

X-original-commit: fd284e3c4a43467bbdd2ca54baa4479c2240c238
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>