[IMP] web_gantt, *: display date range in week scale headers
This commit enhances the display of week group headers in the Gantt view.
In addition to the week number (as before), headers now show the week's
start and end dates in a short format. If the week belongs to a different
year than the current one, the year is also displayed for clarity.

task-4613165

closes odoo/enterprise#85478

Related: odoo/odoo#209901
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>