[REF] ai: decouple translate from editor's chatgpt plugin
This commit removes the translate functionality from the chatgpt plugin
in the ai module.

That is done to effectively decouple the two functionalities since they
are independent in their implementations and their use.

The translate functionality is added back in the html_editor module as a
separate editor plugin.

closes odoo/enterprise#86199

X-original-commit: 4b45856565e13d050f836a52a3a980a15eb537f7
Related: odoo/odoo#211220
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>