[FIX] stock_barcode: barcodeKanban: batch has groups and prevent race condition
In JS, rpc is not a service anymore so it will return no matter if the component is already destroyed.
In turn, the orm service is protected, so if it is passed it will crash.

After this commit, we prevent this race condition and we batch calls that are independant from one another.

runbot-error-190605

closes odoo/enterprise#85673

X-original-commit: 4d1206b06334f534a1f4defd154a12fec838e0b7
Signed-off-by: Steve Van Essche <svs@odoo.com>